requirements.txt
discord.py
httpx
openai
pyyaml
tiktoken
pypng